mv functions common to all launchers to utils
