+++ src/ros_vision_launch/package.xml
@@ -3,11 +3,17 @@
   <name>ros_vision_launch</name>
   <version>0.0.1</version>
   <description>Launch files for vision stack</description>
-  <maintainer email="your@email.com">Your Name</maintainer>
+  <maintainer email="cgpadwick@gmail.com">cpadwick</maintainer>
   <license>MIT</license>
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
 
+  <buildtool_depend>ament_python</buildtool_depend>
+
+  <exec_depend>rclpy</exec_depend>
+  <exec_depend>launch</exec_depend>
+  <exec_depend>launch_ros</exec_depend>
+  <exec_depend>ament_index_python</exec_depend>
+
+  <export>
+    <build_type>ament_python</build_type>
+  </export>
 </package>
